Migrate CI to Pixi (#12389)

+++ dask/dataframe/dask_expr/_collection.py
@@ -2213,7 +2213,7 @@ Expr={expr}"""
         >>> res = ddf.x.reduction(count_greater, aggregate=lambda x: x.sum(),
         ...                       chunk_kwargs={'value': 25})
         >>> res.compute()
-        25
+        np.int64(25)
 
         Aggregate both the sum and count of a Series at the same time:
